Feat: 포토톡 파일 구조 수정
