+++ src/pages/PhotoTalk/AdminPhotoTalkPage.tsx
@@ -1,12 +1,12 @@
-import PhotoTalkList from '@components/display/PhotoTalkSection/PhotoTalkList';
+import PhotoTalkList from '@/pages/PhotoTalk/PhotoTalkList';
 
 const AdminPhotoTalkPage = () => {
   return (
     <div className="result-layout w-full">
-      <div className="column-center w-full">
+      <div className="column-center w-full flex flex-col gap-4">
         <div className="title">포토톡 모아보기</div>
+        <PhotoTalkList isAdmin={true} />
       </div>
-      <PhotoTalkList isAdmin={true} />
     </div>
   );
 };
